Moved service from core to the root.
Refs: #109
